Update pipeline.py
Because try won't fail if youtube-dl doesn't exist.

+++ pipeline.py
@@ -29,17 +29,15 @@ except ImportError:
     print('Please install or update the requests module.')
     sys.exit(1)
     
-try:
-    os.path.isfile('/usr/local/bin/youtube-dl')
+if os.path.isfile('/usr/local/bin/youtube-dl')
     os.symlink('/usr/local/bin/youtube-dl', os.getcwd() + '/youtube-dl')
     print('Created youtube-dl symlink in /data')
-except:
-    try:
-        os.path.isfile('/usr/bin/youtube-dl')
+else:
+    if os.path.isfile('/usr/bin/youtube-dl')
         os.symlink('/usr/bin/youtube-dl', os.getcwd() + '/youtube-dl')
-    print('Created youtube-dl symlink in /data')
-    except:
-        print('Failed to symlink youtube-dl to /data')
+        print('Created youtube-dl symlink in /data')
+else:
+    print('Failed to symlink youtube-dl to /data')
 
 import seesaw
 from seesaw.config import realize, NumberConfigValue
